flask_server/requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.5
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
llama-index==0.10.34
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.34
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.3
llama-index-llms-openai==0.1.16
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.20
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.2
llamaindex-py-client==0.1.19
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.25.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
striprtf==0.0.26
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4